refactor: move buttons into separate component file.
